% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Democrats_State.R
\name{Democrats_State}
\alias{Democrats_State}
\title{Democrats sent by State}
\usage{
Democrats_State(state)
}
\arguments{
\item{state}{Name of the state that you want the information for. Must be capitalized appropriately.}
}
\description{
This function counts the number of Democrats a given state sent to Congress
in each election year from 1998 to 2016.
}
\examples{
Democrats_State("Arizona")
}
