% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Republicans_State.R
\name{Republicans_State}
\alias{Republicans_State}
\title{Republicans sent by State}
\usage{
Republicans_State(state)
}
\arguments{
\item{state}{Name of the state that you want the information for. Must be capitalized appropriately.}
}
\description{
This function counts the number of Republicans a given state sent to congress
in each election year from 1998 to 2016.
}
\examples{
Republicans_State("New York")
}
